Add files via upload
# The Quiz Game

This is my first python project. 
It is a game project for anyone who loves to gain knowledge in a playful manner.
This project asks question to the person playing about computer and stuff and gives them scores and tells the percentage of questions they marked correct.

## 🛠 Skills
Python

## Lessons Learned
In this project I learned to use the indentation correctly, if-else method in python programming language and to print statements while we are adding numbers simultaneously.